v1.0.3 - Strict Security Audit & Clippy Compliance
- Resolved unused imports and unnecessary dependencies across all crates.
- Replaced manual math functions with robust core functions (.div_ceil).
- Repaired unnecessary string formatting in the Lexer.
- Achieved 100% clean check under `cargo clippy --workspace -- -D warnings`.
- Validated all 71 local unit tests passing flawlessly.

+++ Core-Engine/thunder-core/src/transaction.rs
@@ -156,7 +156,7 @@ impl Transaction {
         buf.extend_from_slice(&self.data);
         buf.extend_from_slice(&self.gas_limit.to_le_bytes());
         buf.extend_from_slice(&self.gas_price.to_le_bytes());
-        buf.extend_from_slice(&(self.kind_id() as u8).to_le_bytes());
+        buf.extend_from_slice(&self.kind_id().to_le_bytes());
         buf
     }
